+++ app/src/utils/hooks/journals/helpers/buildMergedData.ts
@@ -1,4 +1,3 @@
-import { journalsRouter } from "src/server/trpc/router/journalEntries";
 import type {
   AccountGroupingReturnType,
   AccountsReturnType,
